requirements.txt
streamlit>=1.47,<2
pandas>=2.2,<3
scikit-learn>=1.5,<2
joblib>=1.4,<2
requests>=2.32,<3
python-dotenv>=1.0,<2
